[Eval GPT-2]: SAE + Masking training for GPT-2

+++ gpt2/main_train.py
@@ -196,7 +196,8 @@ if __name__ == "__main__":
                 ground_truth_token_id = source_label_ids
                 # ground_truth_token_id = base_label_ids
                 vocab_size = model.tokenizer.vocab_size
-                ground_truth_one_hot = F.one_hot(ground_truth_token_id["input_ids"], num_classes=vocab_size).float()
+                ground_truth_one_hot = F.one_hot(ground_truth_token_id["input_ids"], num_classes=vocab_size)
+                ground_truth_indices = ground_truth_indices.to(dtype=torch.long)
                 cloned_intervened_base_output = intervened_base_output.clone()
                 last_token_output = cloned_intervened_base_output[:,-1,:]
                 assert ground_truth_one_hot.squeeze(1).shape == last_token_output.shape
